chore: convert InsightsBookingRepository to service
